Add get_vote_accounts() to RPC client
