+++ client/src/rpc_client.rs
@@ -2,7 +2,7 @@ use crate::client_error::ClientError;
 use crate::generic_rpc_client_request::GenericRpcClientRequest;
 use crate::mock_rpc_client_request::MockRpcClientRequest;
 use crate::rpc_client_request::RpcClientRequest;
-use crate::rpc_request::{RpcEpochInfo, RpcRequest};
+use crate::rpc_request::{RpcEpochInfo, RpcRequest, RpcVoteAccountStatus};
 use bincode::serialize;
 use log::*;
 use serde_json::{json, Value};
@@ -96,6 +96,25 @@ impl RpcClient {
             io::Error::new(
                 io::ErrorKind::Other,
                 format!("GetSlot parse failure: {}", err),
+            )
+        })
+    }
+
+    pub fn get_vote_accounts(&self) -> io::Result<RpcVoteAccountStatus> {
+        let response = self
+            .client
+            .send(&RpcRequest::GetVoteAccounts, None, 0)
+            .map_err(|err| {
+                io::Error::new(
+                    io::ErrorKind::Other,
+                    format!("GetVoteAccounts request failure: {:?}", err),
+                )
+            })?;
+
+        serde_json::from_value(response).map_err(|err| {
+            io::Error::new(
+                io::ErrorKind::Other,
+                format!("GetVoteAccounts parse failure: {}", err),
             )
         })
     }
